+++ src/app/services/album-accent-color/album-accent-color.service.ts
@@ -2,19 +2,20 @@ import { Injectable } from '@angular/core';
 import Vibrant from 'node-vibrant/lib/bundle';
 import { Logger } from '../../common/logger';
 import { MetadataService } from '../metadata/metadata.service';
+import { PathUtils } from '../../common/utils/path-utils';
 
 @Injectable({ providedIn: 'root' })
 export class AlbumAccentColorService {
     public constructor(
         private metadataService: MetadataService,
         private logger: Logger,
-    ) {}
+    ) { }
 
     public async getAlbumAccentColorAsync(albumKey: string): Promise<string> {
         const albumArtworkPath: string = this.metadataService.getAlbumArtworkPath(albumKey);
 
         try {
-            const palette = await Vibrant.from(`file://${albumArtworkPath}`).getPalette();
+            const palette = await Vibrant.from(PathUtils.createFileUrl(albumArtworkPath)).getPalette();
             return palette.Vibrant?.getHex() ?? '';
         } catch (e) {
             this.logger.error(e, 'Could not extract accent color from album cover', 'AlbumAccentColorService', 'getAlbumAccentColorAsync');