+++ src/app/services/album/album-model.spec.ts
@@ -107,7 +107,7 @@ describe('AlbumModel', () => {
             const artworkPath: string = albumModel.artworkPath;
 
             // Assert
-            expect(artworkPath).toEqual('file:////root/directory/dummy');
+            expect(artworkPath).toEqual('file:///root/directory/dummy');
         });
     });